# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Priva Job Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successful

  # Profile
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Company not found
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update company profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # Jobs
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (job seeker sees own, recruiter sees by job)
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: candidateId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status (recruiter)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusChange"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Dashboard
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [job_seeker, recruiter, admin]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    UserRegistration:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [job_seeker, recruiter]
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [job_seeker, recruiter, admin]
        createdAt:
          type: string
          format: date-time

    # Profile schemas
...
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
    Company:
      type: object
      required: [id, recruiterId, name, createdAt]
      properties:
        id:
          type: integer
        recruiterId:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Job schemas
...
    Job:
      type: object
      required: [id, title, description, location, employmentType, workMode, status, postedAt]
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        recruiterId:
          type: integer
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        employmentType:
          type: string
          enum: [full-time, part-time]
        status:
          type: string
          enum: [active, inactive, pending]
        postedAt:
          type: string
          format: date-time
        company:
          $ref: "#/components/schemas/Company"
...
    ApplicationInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
    ApplicationStatusChange:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [applied, viewed, shortlisted, rejected, interview_scheduled, hired]
    Application:
      type: object
      required: [id, jobId, candidateId, status, appliedAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        candidateId:
          type: integer
        status:
          type: string
          enum: [applied, viewed, shortlisted, rejected, interview_scheduled, hired]
        appliedAt:
          type: string
          format: date-time
        job:
          $ref: "#/components/schemas/Job"
        candidate:
          $ref: "#/components/schemas/User"

    # Dashboard schemas
...