# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Priva Job Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  recruiterId: number;
  name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logo?: string | null;
  createdAt: Date;
}